backend/requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-decouple
dj-rest-auth
django-allauth
channels
channels-redis